requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
